MathViz-master/server/requirements.txt
aniso8601==9.0.1
click==8.0.1
cycler==0.10.0
Flask==2.0.1
Flask-RESTful==0.3.9
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
mpmath==1.2.1
numpy==1.21.2
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sympy==1.8
typing-extensions==3.10.0.2
Werkzeug==2.0.1
zipp==3.5.0
